+++ packages/gen-typescript-declarations/src/test/goldens/polymer-3/lib/mixins/properties-changed.d.ts
@@ -59,6 +59,8 @@ interface PropertiesChangedConstructor {
   typeForProperty(name: string): void;
 }
 
+export {PropertiesChangedConstructor};
+
 interface PropertiesChanged {
 
   /**