+++ packages/gen-typescript-declarations/src/test/goldens/polymer-3/lib/mixins/template-stamp.d.ts
@@ -195,6 +195,8 @@ interface TemplateStampConstructor {
   _contentForTemplate(template: HTMLTemplateElement|null): DocumentFragment|null;
 }
 
+export {TemplateStampConstructor};
+
 interface TemplateStamp {
 
   /**