+++ packages/gen-typescript-declarations/CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   particular identifiers are referenced.
 - Automatically detect if a project uses NPM or Bower and configure module
   resolution settings accordingly.
+- Automatically import/export synthetic mixin constructor interfaces.
 
 ## [1.3.0] - 2018-06-29
 - Generate typings for class constructors.